requirements.txt
comtypes==1.1.10
lxml==4.6.4
Pillow==8.4.0
python-pptx==0.6.21
XlsxWriter==3.0.2
